+++ tests/20-EdvDocument.spec.js
@@ -1,6 +1,8 @@
 /*!
  * Copyright (c) 2018-2019 Digital Bazaar, Inc. All rights reserved.
  */
+'use strict';
+
 import {EdvDocument, EdvClient} from '..';
 import mock from './mock.js';
